+++ ModsOfMistriaInstallerLibTests/InstallerVersionTest.cs
@@ -1,0 +1,13 @@
+using Garethp.ModsOfMistriaInstallerLib;
+
+namespace ModsOfMistriaInstallerLibTests;
+
+[TestFixture]
+public class InstallerVersionTest
+{
+    [Test]
+    public void SupportsTheLatestUpstreamMomiManifestVersion()
+    {
+        Assert.That(InstallerVersion.ModCompatibilityVersion, Is.EqualTo(new Version(0, 15, 10)));
+    }
+}